f2fs: remove the ugly pointer conversion
This patch modifies the use of bi_private to remove pointer chasing for sbi.
Previously, we had a bi_private structure, but it needs memory allocation.
So this patch uses bi_private by the sbi pointer and adds a completion pointer
into the sbi.
This can achieve no memory allocation and nice use of the bi_private.

Change-Id: I1546c1e7f991f8eebe9323d6930a5919d18773cb
Signed-off-by: Jaegeuk Kim <jaegeuk.kim@samsung.com>