apexq: tweak torch high brightness
This change does 2 things. First, it restores the previous behavior for
Express that any value above 0 will turn the torch on. Second, it turns
on torch low brightness before turning on high brightness for Apex,
similar to how the flash works when taking a picture. Otherwise there is
a risk of stuck gpio if the phone previously was rebooted with the torch
on.

Change-Id: I9f5c8800dbdd02ca2c21349a6fad9e3f3a7b0042